requirements.txt
absl-py==2.1.0
ace_tools==0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bertopic==0.17.0
bibtexparser==1.4.3
biopython==1.85
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
branca==0.8.1
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
folium==0.19.5
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
fuzzywuzzy==0.18.0
gast==0.6.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gprofiler==1.2.2
gprofiler-official==1.0.0
grpcio==1.71.0
gseapy==1.1.7
h11==0.14.0
h5py==3.13.0
hdbscan==0.8.40
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
ImageHash==4.3.2
ipykernel==6.29.5
ipympl==0.9.7
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.1.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kagglehub==0.3.10
keras==3.9.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lets-plot==4.6.1
libclang==18.1.1
llvmlite==0.44.0
logomaker==0.8.6
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.1.2
mizani==0.13.1
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
multimethod==2.0
murmurhash==1.0.12
namex==0.0.8
narwhals==1.30.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
openai==1.76.0
openai-whisper==20240930
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
palettable==3.3.3
pandas==2.2.3
pandas-profiling==3.2.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
plotnine==0.14.5
praw==7.8.1
prawcore==2.4.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pyogrio==0.10.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypng==0.20220715.0
pyproj==3.7.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
setuptools==76.0.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.1
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.44.1
summarytools==0.3.0
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
update-checker==0.18.0
uri-template==1.3.0
urllib3==2.3.0
visions==0.7.4
wasabi==1.1.3
Wave==0.0.2
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
whisper==1.1.10
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xyzservices==2025.1.0
